+++ notice.go
@@ -20,6 +20,16 @@ type ErrorClass struct {
 	Name string
 }
 
+// Fingerprint represents the fingerprint of the error, which controls grouping
+// in Honeybadger.
+type Fingerprint struct {
+	content string
+}
+
+func (f *Fingerprint) String() string {
+	return f.content
+}
+
 // Notice is a representation of the error which is sent to Honeybadger, and
 // implements the Payload interface.
 type Notice struct {
@@ -36,6 +46,7 @@ type Notice struct {
 	Params       Params
 	CGIData      CGIData
 	URL          string
+	Fingerprint  string
 }
 
 func (n *Notice) asJSON() *hash {
@@ -47,10 +58,11 @@ func (n *Notice) asJSON() *hash {
 			"version": VERSION,
 		},
 		"error": &hash{
-			"token":     n.Token,
-			"message":   n.ErrorMessage,
-			"class":     n.ErrorClass,
-			"backtrace": n.Backtrace,
+			"token":       n.Token,
+			"message":     n.ErrorMessage,
+			"class":       n.ErrorClass,
+			"backtrace":   n.Backtrace,
+			"fingerprint": n.Fingerprint,
 		},
 		"request": &hash{
 			"context":  n.Context,
@@ -150,6 +162,8 @@ func newNotice(config *Configuration, err Error, extra ...interface{}) *Notice {
 			notice.setContext(t)
 		case ErrorClass:
 			notice.ErrorClass = t.Name
+		case Fingerprint:
+			notice.Fingerprint = t.String()
 		case Params:
 			notice.Params = t
 		case CGIData: